+++ Craftsman/Builders/Projects/SharedTestsCsProjBuilder.cs
@@ -31,7 +31,7 @@ public class SharedTestsCsProjBuilder
 
   <ItemGroup>
     <PackageReference Include=""Ardalis.SmartEnum"" Version=""2.1.0"" />
-    <PackageReference Include=""AutoBogus"" Version=""2.13.1"" />
+    <PackageReference Include=""AutoBogusLifesupport"" Version=""2.14.0"" />
     <PackageReference Include=""Bogus"" Version=""34.0.2"" />
   </ItemGroup>