+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Move most `UserPolicyHandler` tests to unit tests
 * Move `CurrentUserServiceTests` test dir
 * Update `LocalConfig` to `Consts` for all constants in app (other than permissions and roles)
+* Update `AutoBogus` to `AutobogusLifesupport` to support .NET 6 and UTC
 
 ### Fixed